requirements.txt
flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1
slack_sdk==3.27.1
requests==2.31.0
gunicorn==22.0.0
